[12169] Fix some problems detected by static code analysis
Conflicts:
	src/game/BattleGround/BattleGroundMgr.cpp
	src/game/ObjectMgr.cpp
	src/game/PathFinder.cpp

+++ src/game/BattleGround/BattleGroundMgr.cpp
@@ -1272,7 +1272,7 @@ void BattleGroundMgr::BuildPvpLogDataPacket(WorldPacket *data, BattleGround *bg)
     {
         ArenaTeam* at[2] = { sObjectMgr.GetArenaTeamById(bg->m_ArenaTeamIds[0]), sObjectMgr.GetArenaTeamById(bg->m_ArenaTeamIds[1]) };
         // it seems this must be according to BG_WINNER_A/H and _NOT_ BG_TEAM_A/H
-        for(int i = 1; i >= 0; --i)
+        for (int8 i = 1; i >= 0; --i)
         {
             uint32 pointsLost = bg->m_ArenaTeamRatingChanges[i] < 0 ? abs(bg->m_ArenaTeamRatingChanges[i]) : 0;
             uint32 pointsGained = bg->m_ArenaTeamRatingChanges[i] > 0 ? bg->m_ArenaTeamRatingChanges[i] : 0;
@@ -1281,7 +1281,8 @@ void BattleGroundMgr::BuildPvpLogDataPacket(WorldPacket *data, BattleGround *bg)
             *data << uint32(bg->GetArenaMatchmakerRatingByIndex(i));    // Matchmaking Value
             DEBUG_LOG("rating change: %d", bg->m_ArenaTeamRatingChanges[i]);
         }
-        for(int i = 1; i >= 0; --i)
+
+        for (int8 i = 1; i >= 0; --i)
         {
             if (at[i])
                 *data << at[i]->GetName();
